termiando las querys
