--- src/app/(pet)/filters/page.tsx
@@ -1,9 +1,0 @@
-import FilterContain from "./ui/FilterContain/FilterContain";
-
-export default function FilterPage(){
-    return(
-        <main className="w-full min-h-screen px-3">
-            <FilterContain />
-        </main>
-    )
-}
