+++ src/app/(pet)/pets/ui/InputFilter/InputFilter.tsx
@@ -3,7 +3,7 @@ import { objectListData } from "@/interfaces";
 /* import { Select, SelectItem, Selection } from "@nextui-org/react"; */
 import { Key, useState } from "react";
 type Props = {
-  listData: objectListData[];
+  listData?: objectListData[];
   name: string;
   setValueOption: /* () => void */ any;
   valueOption: objectState;
@@ -26,10 +26,11 @@ export default function InputFilter({
   /*  const handleSelectionChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     setValue(e.target.value);
   }; */
-  /* console.log(valueOption); */
+  console.log(listData);
 
   const handleChangeValueOption = (e: { target: { value: string } }) => {
     const { value, name }: any = e.target;
+
     let newName = " ";
     if (name === "Sexo") {
       newName = "urlSex";
@@ -58,18 +59,27 @@ export default function InputFilter({
             ? valueOption.urlProvince
             : name === "Edad"
             ? valueOption.urlAge
-            : ""
+            : "cualquiera"
         }
         className="rounded-xl border border-violetGrow-600 p-2"
         onChange={handleChangeValueOption}
       >
-        {listData.map((item: any, index: Key | null | undefined) => {
-          return (
-            <option key={index} value={item.value}>
-              {item.data}
-            </option>
-          );
-        })}
+        {name !== "Edad"
+          ? listData?.map((item: objectListData, index: Key | null | undefined) => {
+              return (
+                <option
+                  key={index}
+                  value={item.slug ? item.slug : item.name.toLowerCase()}
+                >
+                  {item.name}
+                </option>
+              );
+            })
+          : listData?.map((item, index) => (
+              <option key={index} value={item.toString()} >
+                {item.toString()}
+              </option>
+            ))}
       </select>
       {/* <Select
         label="sexo"
